move context menu plugin under the for only directory
