+++ src/plugin/ContextMenuPlugin/ExportJsonDataToClipboardPlugin.js
